+++ OpenCL/MacOSX/ArraySumBenchmark/Stopwatch.h
@@ -4,12 +4,12 @@ class Stopwatch {
     private:
       bool _running;
       long _start;
-      long _time;
+      double _time;
     public:
-	  Stopwatch() { _running=false; _start=0; _time=0;};
+      Stopwatch() { _running=false; _start=0; _time=0;};
       void start();
       void stop();
       void reset();
       void restart();
-      long getTime();
+      double getTime();
 };
